backend/requirements.txt
Flask
Flask-CORS
Pillow
numpy
opencv-python
mediapipe
torch
torchvision
